Move colors get functions
